+++ lib/Braintree/Customer.php
@@ -102,9 +102,9 @@ class Customer extends Base
      * @param string $id customer id
      * @return Customer
      */
-    public static function find($id)
+    public static function find($id, $associationFilterId = null)
     {
-        return Configuration::gateway()->customer()->find($id);
+        return Configuration::gateway()->customer()->find($id, $associationFilterId);
     }
 
     /**